+++ changelog.md
@@ -688,3 +688,6 @@
 - added getSubmissions with sorting and filtering
 - [update1-END]
 - added ability to submit multiple times to a challenge and resume correctly to the "most recent one that is unsubmitted"
+
+### 4/15/24
+- added mode toggle to getChallenge which fixed opening challenge details on submissions page